(11.2) Finaliza formulário do exercício
