+++ 02-desenvolvimento-front-end/bloco-11-componentes-com-estado-eventos-e-formulario-com-react/dia-02-formulario-no-react/exercicios/my-form-2.0/src/StateData.js
@@ -1,0 +1,15 @@
+const INITIAL_STATE = {
+  name: '',
+  email: '',
+  cpf: '',
+  address: '',
+  city: '',
+  state: 'Selecione',
+  addresstype: '',
+  curriculumsummary: '',
+  office: '',
+  descriptionjob: '',
+  submitted: false,
+}
+
+export default INITIAL_STATE;
