+++ 02-desenvolvimento-front-end/bloco-11-componentes-com-estado-eventos-e-formulario-com-react/dia-02-formulario-no-react/exercicios/my-form-2.0/src/components/ButtonCleanAll.js
@@ -1,0 +1,12 @@
+import { Component } from 'react';
+
+export default class ButtonCleanAll extends Component {
+  render() {
+    const cleanForm = () => {
+      window.location.reload();
+    }
+    return (
+      <button type="button" onClick={cleanForm}>Limpar formulário</button>
+    )
+  }
+}
