+++ 02-desenvolvimento-front-end/bloco-11-componentes-com-estado-eventos-e-formulario-com-react/dia-02-formulario-no-react/exercicios/my-form-2.0/src/components/Form.js
@@ -2,6 +2,7 @@ import { Component } from 'react';
 import PersonalInfo from './PersonalInfo';
 import './Form.css';
 import DetailsLastJob from './DetailsLastJob';
+import ButtonCleanAll from './ButtonCleanAll';
 
 
 class Form extends Component {
@@ -13,6 +14,7 @@ class Form extends Component {
           <PersonalInfo handleChange={handleChange} stateData={stateData} />
           <DetailsLastJob handleChange={handleChange} stateData={stateData} />
           <input type='button' onClick={sendForm} value="Enviar" />
+          <ButtonCleanAll { ...this.props}/>
         </form>
       </div>
     )
