+++ 02-desenvolvimento-front-end/bloco-11-componentes-com-estado-eventos-e-formulario-com-react/dia-02-formulario-no-react/exercicios/my-form-2.0/src/components/PersonalInfo.js
@@ -49,7 +49,7 @@ class PersonalInfo extends Component {
         <div className='div-internal'>
           <label htmlFor="state">
             Selecione seu estado:
-            <States statedata={stateData} handleChange={handleChange} />
+            <States { ...this.props } handleChange={handleChange} />
           </label>
         </div>
