+++ 02-desenvolvimento-front-end/bloco-11-componentes-com-estado-eventos-e-formulario-com-react/dia-02-formulario-no-react/exercicios/my-form-2.0/src/components/States.js
@@ -3,12 +3,12 @@ import states from '../data-state'
 
 class States extends Component {
   render() {
-    const { handleChange } = this.props;
+    const { stateData, handleChange } = this.props;
     return (
       <select name="state" id="state" className='text' onChange={handleChange} >
         <option>Selecione</option>
         {states.map(state =>
-          <option key={state.uf} value={state.nome}>{state.nome}</option>
+          <option key={state.uf} value={stateData.state}>{state.nome}</option>
         )}
       </select>
     )
